requirements.txt
gspread==4.0.4
oauth2client==4.1.3
google-auth==2.4.1
streamlit==1.6.0
requests==2.26.0
pandas==1.3.3
plotly==5.3.1